feat: Add Divider component
This commit adds a new `Divider` component to the UI folder. The `Divider` component is a simple horizontal line that can be used to visually separate content. It is implemented as a functional component using React.

The `Divider` component is styled with Tailwind CSS classes to create a 1px high line with a semi-transparent primary background color.

This commit also includes other changes such as adding a new `Project` interface to the `common/types` folder, creating a `NotFound` component, updating the Prisma configuration file, adding a `SmoothScroll` component, creating a `Footer` component, and adding utility functions to the `utils` file.

Note: The recent user commits and repository commits do not provide any additional relevant information for generating the commit message.